fix(eventstore): change offset to uint64
fix(migration): remove u0000 unicodes during copy

fix(migration): use all encryption keys in

+++ internal/eventstore/handler/v2/statement.go
@@ -65,7 +65,7 @@ type Statement struct {
 	CreationDate  time.Time
 	InstanceID    string
 
-	offset uint16
+	offset uint64
 
 	Execute Exec
 }
